+++ wireform-kafka/test/Main.hs
@@ -65,14 +65,14 @@ import qualified Client.ResponseFrameSpec
 import qualified Network.OAuthOidcSpec
 import qualified Network.BootstrapSpec
 import qualified Network.ReauthDriverSpec
-import qualified Client.ConsumerExtrasSpec
+import qualified Client.ConsumerSnapshotsSpec
 import qualified Client.MetadataCacheControlSpec
 import qualified Client.RackAwareSpec
-import qualified Client.ProducerExtrasSpec
-import qualified Client.ConnectionExtrasSpec
-import qualified Client.AdminExtrasSpec
+import qualified Client.TransactionHelpersSpec
+import qualified Network.ConnectionHelpersSpec
+import qualified Client.AdminClientConfigSpec
 import qualified Client.SerdeContextSpec
-import qualified Client.ShareGroupExtrasSpec
+import qualified Client.ShareConsumerHelpersSpec
 import qualified Client.AdminClientSpec
 import qualified Client.MockBrokerSpec
 import qualified Client.MockBrokerFailureModesSpec
@@ -172,14 +172,14 @@ clientTests = testGroup "Client"
   , Client.RetryClassifierSpec.tests
   , Client.BatchSplittingSpec.tests
   , Client.ResponseFrameSpec.tests
-  , Client.ConsumerExtrasSpec.tests
+  , Client.ConsumerSnapshotsSpec.tests
   , Client.MetadataCacheControlSpec.tests
   , Client.RackAwareSpec.tests
-  , Client.ProducerExtrasSpec.tests
-  , Client.ConnectionExtrasSpec.tests
-  , Client.AdminExtrasSpec.tests
+  , Client.TransactionHelpersSpec.tests
+  , Network.ConnectionHelpersSpec.tests
+  , Client.AdminClientConfigSpec.tests
   , Client.SerdeContextSpec.tests
-  , Client.ShareGroupExtrasSpec.tests
+  , Client.ShareConsumerHelpersSpec.tests
   , Client.AdminClientSpec.tests
   , Client.GroupSpec.groupSpec
   , Client.MockBrokerSpec.tests